[Tags] Undo a few post processing related values and blocks
Vehicle
If gravity scale is set to 0 then it is turned into a 1 on package. This sets gravity scale on extracted vehicles to 0.001 which should turn into a 0 on package instead of a 1

Bitmap
Cleared a byte in the bitmap flags. Should hopefully let bitmaps work out of the box.

Physics Models
Deletes the phantom tag block. This block is only created on package.

Sets phantom enum in physics mesh tag blocks to none. This is set by the referenced phantom mesh in the material on package.