[UI] Make extraction run on a new thread and some misc changes
Extraction now runs on a new thread. This means that the program no longer stops responsing and can be messed with during the process.
This has brought out some new problems such as

Attempting to load a resource map after loading a regular map file will result in a crash due the file being in use. I assume this has something to do with the asserts. This bug existed before but turned from the program refusing to load resource maps to crashing while attempting to load them.

The "selected tags" count is cleared when loading or closing a map.

Close the map_stream while loading and closing the map to avoid some  crashes.

Disable parts of the UI related to extraction during extraction to avoid some crashes.

A really trash workaround for the whole "can't extract from shared properly thing". Isn't the best thing that could probably be done but this solves the issue for now.